renamed tcb_info pod module
